VersionContext does not implement GraphQL Node
The GraphQL Node ID needs to be able to be looked up globally. But the VersionContext ID was just the string name of the version context, which is not globally addressable. Plus, there is no switch-clause in the `Node` resolver impl to look up a version context anyway.

There are no places in the frontend code that attempted to look up a VersionContext by its node ID. If there were any, they would not have worked anyway. So, this is safe to remove.